Be able to disable the file interface.
Found a case where the file is created by a user, and fgms is run as
another user, thus is unable to delete the file. Thus need to be able to
disable that particular file interface so that file will not be repeatedly
found. It will stay disabled for that particular fgms session.

Of course this would NOT happen if fgms is run as root, but that should be
the users choice.

As a consequence this also gives a way to disable this file interface at
start-up, just by creating such a file before starting fgms.

Bumped version to 0.11.9 for this relatively minor file interface change.