Issue #6 - Add a euler_get() service to SGUtils lib.
This was formerly sgdebug library, but with the addition of the new
euler_get() service rename to SGUtils.

The addition of math/SGEuler.[ch]xx allows the telnet orientation XYZ to
be converted to heading, pitch and roll.

An example usuage has been added to the server so the POSITION tracker
message can use heading, pitch, roll...