Try to improve on previous usleep for WIN32, and add alternative
Issue #11 - Improve on previous implemetation, actually sleeping if there
are whole ms involved, and only burn CPU time for any microsecond
remainder...

Also add an alternative using WaitForSingleObject on a timer handle. This
can be used by defining USE_WAIT_OBJECT.

Of course this is ONLY for WIN32, using the system usleep in linux.

In tests, be aware the actual sleep starts to not be very precise below
about 10 milliseconds... and will usually be less than the requested...